Use the asynchronous interface for setting the default_encoding.
This is another attempt to solve issue https://bitbucket.org/ged/ruby-pg/issues/245
While setting the encoding per PQsetClientEncoding was already adjusted to
release the GVL in a previous commit, it did not solve the particular issue
with pgbouncer.

Using the async interface has the advantage, that signals or exception events
can be processed by the given thread, while a query is running.