MINGW: Workaround segfault due to gcc linker error
Linking to a MSVC compiled libpq.dll can lead to segfaults due
to incorrect address offsets of function calls.
This happens on PostgreSQL-10.0-x64 from EnterpriseDB.
It can be avoided by linking to the import library (libpq.lib)
instead of the libpq.dll.

Unfortunately linking to the import library fails when running an older
gcc (4.7.2), so that we check the gcc version as well.

The issue is filed upstream (for gcc-7.2.0):
    https://sourceware.org/bugzilla/show_bug.cgi?id=22504

I didn't check any gcc version between 4.7.2 and 7.2.0, but other
gcc versions are uncommon on Ruby on Windows. So the check is
for gcc version >= 6 for now.