Add a bunch of new specs for timestamp decoders
These specs are extracted out of the following pull request:
https://github.com/ged/ruby-pg/pull/19
which has the intention is to speed up SimpleDecoder::TimestampWithTimeZone and
SimpleDecoder::TimestampWithoutTimeZone .

Since the purpose of the PR is performance only, we can safely import
the additional specs and ensure that the current RegExp based parser or any
future parser deliver the same results.

Specs are written by @remy-cassia-tech and adjusted by me.