Add PG::Connection#discard_results and use it in async_exec
See PostgreSQL source file fe-exec.c method PQexecStart().

Previously we discarded results per get_last_result, but this didn't
clear a COPY IN/OUT state. This is now fixed, so that async_exec
has the same behaviour as exec in the COPY context, now.

This also removes a call to pgconn_block(), which was required
on Ruby-1.8 to not block other threads while discarding results.
Now the GVL is released anyways. In case PQgetResult() blocks,
it will now block signal processing, but since discarding
results is only a fallback for mis-usage of the pg API, we
save this call for better performance.