Make exec and siblings switchable between sync and async API
Most applications don't care about the subtile differences between #exec
and #async_exec , but in general #async_exec is recommended, since #exec
blocks signal processing for the time of the running query.

This commit renames the classic synchronous methods to sync_* .
Furthermore it links #exec and siblings to the new async_* method versions.
This makes the use of the new async methods easy, without the need to chance
any calling code.

Since we run the PG::Connection specs on both APIs (sync_* and async_*)
we make sure, that both versions behave identical in all use cases.