Revert parsing of malformed array strings and deprecate it
With commits
  https://bitbucket.org/ged/ruby-pg/commits/0567662910977452ad8b9b5a6eb67bd5dd8e67ec
and
  https://bitbucket.org/ged/ruby-pg/commits/6a6416e72e933199e2dbc8f5d9d7fd6b383bba61
I changed TextDecoder::Array to raise a TypeError for malformed strings.

Previously we expected Arrays (without dimensions only) according to the PostgreSQL specs
  https://www.postgresql.org/docs/current/static/arrays.html#ARRAYS-IO
and returned a more or less useful ruby Array in any case, even when the array
string was malformed.

The reason to better raise a TypeError is, that there's no definition of how
malformed array strings shall be parsed. This is because they are outside of the
specification...

Unfortunatelly rails makes use of the feature to parse malformed arrays here
  https://github.com/rails/rails/blob/v5.2.0/activerecord/lib/active_record/connection_adapters/postgresql/oid/array.rb#L36
and tests for it here
  https://github.com/rails/rails/blob/v5.2.0/activerecord/test/cases/adapters/postgresql/array_test.rb#L316-L330

Therefore I think it's best to re-enable this behaviour for now and deprecate it.
We can switch to raising TypeErrors with pg-2.0. This is prepared with this commit.