Fix various issues in the inet/cidr decoder
I found these issues by porting the code to sequel_pg and running
Sequel's tests.

It turns out that IPAddr doesn't support non-zero parts after the
netmask, such as 1.2.3.4/24, and automatically modifies the part
after the netmask to be 0 (turning 1.2.3.4/24 to 1.2.3.0/24).
Because the default is to set instance variables for performance,
this resulted in the creation of IPAddr instances that you could
not generate using IPAddr's public API.  To fix that, mask the
part of the IP address integer(s) before calling UINT2NUM or
ULL2NUM.

I've modified the inet tests to be much simpler, just checking
that the returned value for the inet column matches what you
would get if you parsed the input string with IPAddr.new. I've
also added tests for the above issue.

Modifying the passed in char* buffer is a bad idea, as that
modifies the underlying PG::Result storage directly, leading to
multiple calls to getvalue returning different data.  Instead,
copy the input string to a stack buffer work off of that if
there is a netmask.  This uses a buffer size of 64, but I think
the maximum for IPv6 is only 40 if you count the trailing string
terminator.

This also calls rb_global_variable for s_IPAddr so that you don't
get undefined behavior if the IPAddr constant is garbage collected.