Change LazyTuple to just Tuple and add name based access.
PG::Tuple now supports access by index and by column name.
This makes it more usable in ActiveRecord, which does all
lookups through column names as String.

This could be extended to PG::DetachedTuple in the future,
to copy all data out of the result to build a detached tuple.