Reject string with zero bytes in quote_ident
PostgreSQL doesn't support zero bytes in identifiers, but uses
them as termination character.
Libpq's quoting functions truncate strings at the first zero byte.
This is no problem, as long as the calling application equally uses
zero terminated strings. But Ruby's strings support zero bytes,
so that the strings are truncated at the first zero byte.

Truncating the string means a different string is sent the server
than intended by the application.
This can also have a security implication to string validations.

We already reject query parameters with zero bytes, so that it's
consistent to deny them in quoting functions as well.

Simplify and optimize the code a bit while being over it.