Revert "Move PQsocket() to connection init for better performance"
Storing a copy of the socket file descriptor has the risk of
getting out of sync with the socket stored in PGconn. This
could happen while conn.reset().

Instead of fixing this, I propose to revert this commit.
The performance benefit is not mesurable and it is in a code section
where the client is waiting for the server anyways.

Fixes #282 (hopefully)