Update versions of Ruby and PostgreSQL running on travis-ci and appveyor
We have to change the build of PostgreSQL to be installed on Linux.
This is because the EDB Linux installer is discontinoused:
https://postgresrocks.enterprisedb.com/t5/Postgres-Gems/Platform-Native-EDB-Packages-for-Linux-Users/ba-p/2362?_ga=2.183359910.1234512303.1545833962-846499147.1545833962

Luckily apt.postgresql.org provides a way to install older versions of libpq.
This is important to verify that we still can build and link to these older versions.

Also generate a dummy RSA key and cert for signing, since I couldn't find
an easy way to disable gem signing.