Define encode and decode methods at the class level, where they are implemented
Previously all Coder classes had encode and decode methods defined, but they raised
an RuntimeError unless a underlying C function has been assigned.
This makes it hard to query for the capabilities of a given coder.
It also doesn't follow the ruby-pg rule, to define only those methods, that are actually usable.

This patch moves the definition of encode and decode to where the C function is assigned,
so that respond_to? can be used to detect whether it's an Encoder or Decoder.

Also use this capability to enrich the output of TypeMapByColumn#inspect .