Assign random port to consider a case multi tests in parallel on a host.
There is a case for the failed test on a build system.

There are some ruby-pg directories in same host
to build and run test on different arch condition.
Then those directories' tests are executed in parallel.

For the case, tmp_test_specs/setup.log shows below error
when starting PostgreSQL from spec/helpers.rb.
Because the tests are using same testing port for each other.

waiting for server to start....2019-05-13 16:25:07.565 UTC [2106] LOG:
 could not bind IPv6 address "::1": Address already in use
2019-05-13 16:25:07.565 UTC [2106] HINT:  Is another postmaster
already running on port 54321? If not, wait a few seconds and retry.

Assigning random port fixes the situation's issue.