Update sync_* and async_* query method descriptions
Since pg-1.1 the async_* variants are the primary and sync_* are the secondary.
To reflect this fact in the documentation the descriptions are now swapped as well.
This also adjusts the aliases, so that the documentation is properly linked.
And this adds links to libpq's corresponding documentation, to make some more details available.