Add ability to use symbols as field names instead of strings
This adds:
* PG::Result#field_name_type=
* PG::Result#field_name_type
* PG::Result#field_names_as

Symbol comparison and lookup is faster, since the string hash is generated at Symbol creation.
So retrieval of Symbol based field names is somewhat slower, but this is easily outperformed in frameworks that do a lot of lookups.

This raises ruby version requirement to 2.2, since it depends on GC'able symbols.

Fixes #278