Fix and improve PG::Result#stream_each_tuple
This fixes a double free of pgresult which happend when the block of stream_each_tuple raised an exception.
That is prevented now by setting pgresult to NULL before the rb_yield call.

This commit also fixes typemap use in combination with stream_each_tuple.
Before the typemap of the connection was used in PG::Tuple instead of the typemap of the base result.

While being at it, I also improved the performance of stream_each_tuple.
The per-tuple PG::Result is now copied per memcpy instead of being initialized based on the connection.
It also shares field names between all PG::Tuple objects now, instead of repeated allocation of the same values.