Add Time object conversion to BasicTypeMapForQueries
This adds an encoder for Time objects to input class based value conversion.
The encoder does the same as the previous Time#to_s conversion, but passes subsecond precision to the server.
The timezone is passed to the server and no time zone conversion is done.

Time object conversion is already done by BasicTypeMapForResults and BasicTypeMapBasedOnResult by the same encoder, so that it's consistent to do the same by the class based type map.