Update to rake-compiler-dock-1.0
Version 1.0 switches to per-target docker images.
So there is a specific image for x64-mingw32 and one for x86-mingw32.
However our current logic mixes the build process of x86 and x64 behind several non-platform specific rake tasks.
This fails with rcd-1.0 since each image has only x64 OR x86 tools.

The solution in this commit is to separate each task by the platform value and to reference only the needed platform.

As an advantage we can now execute docker builds in parallel for x64 and x86 per multitask.
Since this conflicts with interactive password prompts, the gem signing key can be entered in advance.

rcd-1.0 does stripping of binaries per default, so our manual tasks are needless.