Relax timing dependent test about timeout calculation
This test was introduced in commit 241ab211cab1f95a8626abc5ef1fa483a9c9d222 to verify select()-timeout calculation in methods waiting for server response.
The test sometimes failed on very slow computers on MRI, but quite often on TruffleRuby.

Unfortunately there is no ultimately reliable way to verify the timeout calculation.
We can not use mocks since the calculation is in C code.
And the impact is time related only.

So this attempt retries the test when expectations fail which are runtime dependent.
It also describes what the checks are for, to allow better understanding of their purpose.

Fixes #331