Replace deprecated rb_cData class by rb_cObject
rb_cData is an alias for rb_cObject since ruby-3.0 and will be removed in ruby-3.1.
Until ruby-2.7 differ in that way that rb_cData doesn't define an allocation function.
With ruby-3.0 it's necessary to undefine it, in order to avoid segfaults with objects returned from PG::Result.allocate or .new .