Make BasicTypeRegistry a class and use a global default instance
Until now only one global type registry was available.
This could lead to issues, if different parts of an application register competitive type coders.
To avoid global shared states, now a local type registry can be instanciated and given to the type map.

A registry can optionally be populated by define_default_types with predefined coders.

This also changes BasicTypeMapForQueries to not raise an exception on undefined types, but emit a warning only.
This makes it usable with partly populated registries.

The new local type registries can be used in the specs, so that it's unnecessary to restore the global state.
And this adds specs for registration of coders in a type registry.