Disable GVL unlock/lock mechanism
Since we do all blocking operations through ruby mechanisms now, we don't need to unlock the GVL any longer.

GVL unlocking was introduced in pg-0.15.0 as a way to allow other ruby threads to be executed while blocking operations.
Unfortunately it turned out that the blocked ruby thread doesn't react to signals and that this method is incompatible to Fiber.scheduler.

This doesn't remove the unlock/lock mechanism from the sources yet, to allow easy comparison.
My plan is to remove the code entirely before the next release.