TCP-Gate: Improve write handling
Write until the observed fd gets writeable again and abort at this point in time.
This is necessary to reliable trigger the next writable-event.
The amount of data to read from internal_io until a writeable state is triggered on the other end
increases with the number of bytes written.

This way we make sure, that only the minimum number of bytes is written, that is necessary to get the nonblocking mechanisms working.