Move async_exec methods back to C
This moves async_exec back to C and enabled other query methods like async_exec_prepared to be fully scheduler compatible.
This also removes async/sync_send_* method variations in favor of using a flag to enable this.

The performance of all query methods should now be similar to pg-1.2.