Allow specification of multiple PostgreSQL hosts to connect to
Unfortunately the uri gem from stdlib isn't compatible with postgresql://host1:port1,host2:port2 syntax.
Therefore we introduce an extended regexp for parsing the URI.

Previously we allowed changes to the URI by overwriting URI parts with hash values.
This is no longer possible and this is a specification change.
But I think it's bad practice and a questionable feature.

Now only additions to the passed URI or connection string are possible.
That way hostaddr and fallback_application_name are automatically appended.

This patch also expands address resolution for multiple hosts.

Fixes #387