Add async_cancel as a nonblocking version of conn#cancel
There's no nonblocking version of PQcancel(), so that we have to implement our own, to make it scheduler compatible.
There's PQbackendPID() to retrieve the necessary process ID from libpq, but unfortunately there is no public function to retrieve the cancel key.
Therefore we mimic the C struct pg_cancel of libpq in order to access the key from the opaque pointer.
It is stable since 17 years, so that it's not expected to change soon.

"be_key" is behind "salen" in "struct SockAddr", which has a size that depend on the outcome of the PostgreSQL configure script.
We define it as "socklen_t" for now. Let's see how well this works on different platforms.

To implement our own cancel method, the new method conn.backend_key is added.

And this changes conn.socket_io to a BasicSocket instead of an IO.
This allows to call methods like remote_address without any "for_fd" re-wrapping.