Replace Resolv.getaddress by IPSocket to avoid inconsistencies
The resolv library resolves differently than the system resolver.
In general it prefers IPv4 over IPv6.
This leads to the situation, that the server socket of the TcpGateScheduler is bound to IPv6, but PG.connect tries to use IPv4 instead.

This happened on Windows-10 and Windows-Server 2016 with Ruby-3.0 and with no entries in the /etc/hosts file.

I tried to align the results of resolv and system library, but didn't find out how the system resolver decides between IPv4 and IPv6.

So the workaround is to use a second thread and use the system resolver on Ruby-3.0.