Fix IPv6 address parsing in connection URI
It's not correct to split host and port by just looking for ":", since IPv6 addresses contain this character.
So this commit uses the host-and-port Regexp to grap the relevant data.

In contrast split(",") is OK to separate hostnames, since the hostname might not contain a comma.
But use split(",", -1) to retrieve a possible last empty hostname.

Also fix interpretation of percent-escapes used for UnixSocket paths in URIs.