Fix sporadic Errno::ENOTSOCK errors on Windows
.. and close socket_io in case of failures.

PQconnectPoll closes the connection in case of failures.
This can lead to a re-use of the socket HANDLE while the ruby IO is still active.
So in this case two distinct file descriptors point to one and the same HANDLE leading to ENOTSOCK errors.
Unwrapping the socket immediately after a socket failure fixes this.

Fixes #398

See more description in https://github.com/ged/ruby-pg/issues/398