Improve API documentation and C function names
Now all methods with async/sync variants have the same form:
The documentation is on the base method, which is documented to have an alias to the async-version.
The corresponding sync method has no documentation or just a note about the sync vs. async difference.

Also add some aliases of Connection.new to async_api switch.