Remove FL_TEST and FL_SET
Object flags are pretty internal information so we shouldn't be checking
them (if possible).  In this case, we don't need to check the status of
EXIVAR because `rb_copy_generic_ivar` will check the flag and return if
there are no ivars to copy[1] and will set the flag on the cloned object
if it did copy[2]

Just for some background, `FL_EXIVAR` flag indicates that an object has
instance variables, but only objects that are not of the type
`T_OBJECT`, `T_CLASS`, or `T_MODULE`.  This type of knowledge is really an
implementation detail, so I think C extensions should not know it or use
it.

1. https://github.com/ruby/ruby/blob/24909f5b3de10b6f0b6f9acbf44ed347ccabb7e7/variable.c#L1725-L1727
2. https://github.com/ruby/ruby/blob/24909f5b3de10b6f0b6f9acbf44ed347ccabb7e7/variable.c#L1741