Fix for TruffleRuby, raise original exception explicitly
Previously this test failed on TruffleRuby because the finally raised exception
was a RuntimeError with no message. It looks like TruffleRuby doesn't raise the
originally caught exception in this case, ie if there is an intermediate
exception between the original one and the raise with no args. Fix by explicitly
raising the originally caught exception.