Don't flush at each put_copy_data call, but flush at get_result
This better mimics, what libpq does internally.

put_copy_data is significantly faster when it doesn't flush at every call.
This is by a factor of 4 on Linux and 10 on Windows when sending typical per-row blocks of 60 byte.

put_copy_end unconditionally calls flush in libpq, so it is not changed here.

Also adjust conn.block to send all enqueued data to mimic the behavior of get_result in libpq.
With the change to put_copy_data, unsent data can happen, when largs blocks are sent.
In this case get_result should catch up on flush.
This is what the newly added spec verifies.

PQgetResult does flushing based on it's internal states that we don't have access to.
Since conn.block is performance critical in case of single_row_mode, we don't flush at every call to conn.block, but only when it's about to wait for IO.