Replace send_query by send_query_params in pipeline mode
Support for PQsendQuery() has been removed from libpq because it was buggy:
  https://github.com/postgres/postgres/commit/0032a5456708811ca95bd80a538f4fb72ad0dd20

These affected our pipeline tests as well, so that we had blocking behaviour on PostgreSQL-14.
This was because sync events were not delivered to pg, so that we received less PGRES_PIPELINE_SYNC than calls to pipeline_sync were done.