Add PG::TextEncoder::Bytea to BasicTypeRegistry
This encoder used to be excluded from the registry, since it's the only encoder that has a negative performance implication.
However it should be there for completeness and because it's valueable for sending data per COPY command to the server.

Another issue was, that we supported PostgreSQL-8.4 to the time TextEncoder::Bytea was implemented, but the hex bytea format was introduced with PostgreSQL-9.0.
Now we don't support such old versions any longer, so that the encoder can be used bar none.