Improve error message when query is canceled during streaming
When streaming a result in single row mode, if this query is aborted -
either because it timed out or some other reason -, `ruby-pg` raises the
error:

```
PG::InvalidChangeOfResultFields: number of fields changed in single row
mode from X to 0 - this is a sign for intersection with another query
```

This error message is misleading and might confuse people.

This commit adds a different error message, using the same exception
class, explaining that the query was canceled or timed out.