Move nfields-check of stream-methods after result status check
This ensures that the nfield-check doesn't hide errors like statement timeout.

Now the new PGresult is assigned to the streaming PG::Result, even if it has a different number of fields.
This avoids leaking the memory of the wrong PGresult, but would lead to wrong fields or segfaults, if the PG::Result is used afterwards.
Therefore the result is getting cleared before the raise.

Thanks to Neil Carvalho @neilvcarvalho for the test case!

Fixes #507