Remove the encoding cache
pg_encoding_to_char() is index based and rb_enc_find() is st_table based, so they are fast.
In addition move the dominant UTF-8 encoding to the very first in the encoding map, so that no loop is necessary.

This cache was only used at connection setup and when calling encoding getter/setter, so that no practical performance impact is expected.

Removing the cache makes connection setup thread safe.
This is required to make ruby-pg ractor ready.