Fix possible segfault when creating new PG::Result
Initialize connection and typemap prior to any object allocations, to make sure valid objects are marked.

This regression was introduced in commit 5061020c28d694464a5fa5474062d8486912daa1 while introducing write barriers.
However it is not necessary to use RB_OBJ_WRITE when the "old" object (1st argument) is not yet created or immediately after it was created.
The initial assignment can and must be done before processing the typemap as it was before the above commit.

Fixes #530