Make `pgconn_connect_poll` close the socket prior to calling `PQconnectPoll`
Since `PQconnectPoll` can change the underlying socket and thus the file
descriptor, closing the socket after can manifest a race condition
where libpq has closed the socket and another thread has opened a socket
with the recycled file descriptor. When the original socket is closed
from Ruby, the VM will notify the new thread that the FD has closed
(erroneously) resulting in an exception.

The same for `PQresetPoll`.