Initial pass adding AWS IAM Authentication
This adds AWS IAM Authentication, which replaces the `password`
parameter normally provided with a dynamically generated, short-lived
authetication token. The authentication token (password) will be
fetched from IAM and cached for the next 14 minutes (tokens expire in
15 minutes).  These can then be reused by all new connections until it
expires, at which point a new token will be fetched when next needed.

To allow for multiple configurations to multiple servers, the cache is
keyed by database username, host name, port, and region.

To use, specify in the environment variables:
- PG_USE_IAM_AUTHENTICATION=true
- AWS_REGION=us-east-1  or whatever your AWS region is

As prerequisites, you must enable IAM authentication on the RDS
instance, create an IAM policy, attach the policy to the target IAM
user or role, create the database user set to use the AWS
Authentication Plugin, and then run your ruby code using that user or
role.  See
https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.IAMDBAuth.Connecting.html
for details on these steps.

You must include the aws-sdk-rds gem in your bundle to use this
feature.  A LoadError is raised if the functionality is tried to be
used but that gem is not present.

The use of environment variables to invoke this functionality is
unusual. You'd expect to able to put these on the postgres connection
string.  However, only arguments valid to PostgreSQL itself are
allowed on the connection string, and only this gem needs them.  If
there's a better way, please advise.