Fix writing lots of data through the TcpGateScheduler
3 tests failed after upgrade to Linux-6.10 due to starvation of data transfer in write direction.
This was because we stopped the transfer when less than the requested size could be read.
Stopping only when no data at all was read fixes this issue.

On Windows reading after partial data read can lead to Errno::ECONNABORTED, so this is considered a EOF condition.

It also turned out, that the same issue led to starvation of SSL encrypted traffix.
So this patch re-enables SSL for scheduler tests which was disabled in commit f270b714c66690b18a0e28bf0af667212cb00bee

Fixes #601