Add support for new query cancel functions of PostgreSQL-17
This adds the new class `PG::CancelConnection` which provides the ability to cancel a query per blocking or per non-blocking functions.
If the new functions are available they are used and the older no longer compiled in.

This way we can get rid of reading out the internal `PGcancel` struct by `Connection#backend_key` .