Use a connection encoding compatible to "Möhre"
The function PQescapeString was changed in PostgreSQL 17.4, 16.8, 15.12, 14.17, and 13.20.
It now returns a predefined invalid character as a replacement in the connection encoding, if the input text is not valid according to the current connection encoding.
Using a compatible connection encoding avoids this, so that we get the original text out of the singleton escape function.

Fixes #628